+++ backend/app/store/user/user.go
@@ -1,10 +1,12 @@
-package engine
+// Package user provides implementations for Interface for the database user repository.
+// All consumers should be using service.DataStore and not the naked repositories!
+package user
 
 import (
 	"github.com/yaattc/automatic-time-table-creation/backend/app/store"
 )
 
-// Interface defines methods to store, and fetch models
+// Interface defines methods to repository, and fetch models
 type Interface interface {
 	GetUser(id string) (u store.User, err error)
 	GetPasswordHash(email string) (pwd string, err error)
